Shrink heap during compaction
Summary:
Remove the compactee segment from the OG during compaction. This lets
us shrink the heap over time. The segment is stored in
`compactee_.segment` until it is fully evacuated, and then it is
destroyed either after evacuation or after sweeping, whichever is
later.

This requires adding some additional code to make sure that the
separate compactee segment is still tracked for things like iterating
over the heap.

Reviewed By: dulinriley

Differential Revision: D24715309

fbshipit-source-id: bc403b69480f89a0191ec681f2aed651c5298bc3