Use the snapshot ID as a key in AllocationLocationTracker::stackMap_
Summary:
After shrinking the node ID to be a `uint32_t`, this will reduce the size
of the stack map.
It also means there's no need to call `moveAlloc`, since the ID is stable
even when the object moves.
Accessing the ID from the pointer isn't a performance issue,
since it's needed anyway in both the `newAlloc` and `freeAlloc` paths.

Reviewed By: neildhar

Differential Revision: D25668990

fbshipit-source-id: b622fbffdd0ff91042bf9219e287373dfa178423