Use compressed pointers in IDTracker::objectIDMap_
Summary:
To further reduce the amount of memory used by `objectIDMap_`,
change the key from a 64-bit value to a 32-bit value by compressing
the pointers. This only applies to a 64-bit system, any build without
compressed pointers will be unaffected.

In practical terms, this reduces the map size by 25%. When the map
can often grow to be 10 MB large, it adds up to significant savings.
The only runtime cost is extra compression on debugging or profiler
paths.

Reviewed By: neildhar

Differential Revision: D25671141

fbshipit-source-id: 13f4fc3aff027f50f84e4483b88e1ba9040bd94d