Simplify generator detection in lazy IRGen.
Summary:
We know after reparsing a lazy function whether it's a generator.
Use that information when performing IRGen, and avoid sending
the `isGenerator` flag at all.

This cleans up generator handling in lazy mode and provides an example
of how to do specialized function IRGen in the future in lazy mode,
e.g. for async functions.

Reviewed By: Huxpro

Differential Revision: D25685446

fbshipit-source-id: ded7dbe59a90fb3b6e7016adc534d4b24d688a2e