Consistent numbering for capture groups
Summary:
The way we handle the bytecode for capture group indices is currently
confusing. There are two types of instructions that store them.
1. The `MarkedSubexpression` instructions themselves. These store a
`mexp` value.
2. Instructions that need to track the number of capture groups they
contain, like loops and lookarounds. These store a `mexpBegin` and
`mexpEnd` to denote that they contain all the marked subexpressions
from `mexpBegin` up to (but not including) `mexpEnd`.

The confusing piece, is that the numbering the two types of
instructions use is off by one. The `MarkedSubexpression`
(and `BackRef`) instructions essentially start at 1. Due to a non-obvious
consequence of how the code is written, the other types of instructions
implicitly start at zero.

So, for instance, a `Loop` that contains `MarkedSubexpression`s with
indices 3, 4, and 5 would have `mexpBegin = 2` and `mexpEnd = 5`.

I initially thought this was a bug, but trying to fiddle with capture
groups inside loops didn't yield any bugs. I then realised that the
off-by-one error is reconciled at runtime, by subtracting 1 from the
index in `MarkedSubexpression` instructions.

This diff makes a small change to make the numbering consistent.

Since we have to update the bytecode version, I also updated the stored
`mexp` values in the bytecode to be 16 bit instead of 32 bit.

Reviewed By: avp

Differential Revision: D25237619

fbshipit-source-id: 2b373d609edd87af30c58f11a1e995ce361fa2d5