Add raw property to ESTree Literal nodes
Summary:
While not officially in the ESTree spec, ESTree Literal nodes often have a `raw` property that contains the raw source code for the literal. Babel includes this property when it is configured to output an ESTree AST, and ESLint expects that all Literal nodes contain this `raw` property.

The Hermes AST does not contain a raw property, but luckily it is still straightforward to add to the ESTree adapter since all we need to do is pull out a substring of the code based on the literal's range (since the range are indices into the source code string).

Reviewed By: tmikov

Differential Revision: D25619522

fbshipit-source-id: 21c714fcc139169f8ccdda85c24c68145d663e85