Generate ESLint visitor keys
Summary:
Since we are supplying a custom parser to ESLint, we should also supply custom visitor keys to specify how the AST can be traversed (https://eslint.org/docs/developer-guide/working-with-custom-parsers). The visitor keys should be a map from node name to an array of property names that can be traversed as strings.

Since the Hermes AST is is based off ESTree, we can generate these visitor keys directly from the Hermes AST definition using the C preprocessor, just like the parser's visitor keys are generated used in the adapters. We can create a template file that contains the Hermes AST macros used to create arrays of child properties that should be visited. Since only node and array children should be visisted, we can define only the `NodePtr` and `NodeList` macros (leaving `NodeLabel`, `NodeBoolean`, and `NodeNumber` blank) so that only node and node list child property names are included in the visitor keys.

However the Hermes AST diverges slightly from the ESTree spec. In particular the Hermes AST defines specific literal nodes for strings, booleans, numbers, etc, while ESTree defines a single `Literal` node. We must be sure to manually add an entry for `Literal` nodes in the visitor keys map. In addition we must also supply an entry for `ThisTypeAnnotations` as like `Literal` nodes they are produced by the ESTree adapter but not defined in the Hermes AST definition.

We need to generate separate visitor keys than the parser uses because A) the visitor keys are in a slightly different format, and B) a slightly different set of visitor keys should be produced. This diff renames the parser visitor keys to `HermesParserVisitorKeys.js`, whereas the ESLint visitor keys are named `HermesESLintVisitorKeys`.

Reviewed By: gkz

Differential Revision: D25401324

fbshipit-source-id: 610879da3ace35bf2da6e2b3c81b657273dfb6e7