Simplify collection check in creditExternalMemory
Summary:
We technically need a lock to access OG external bytes and target
bytes, which we currently do not do.

In addition, the current YG external memory check, which checks the
OG collection criteria, is generating a lot of spurious collections,
particularly since it does not check whether we are already in an OG
collection.

Instead, simplify all of these criteria to be as permissive as
possible, since external memory is rare.

Reviewed By: dulinriley

Differential Revision: D25657260

fbshipit-source-id: 7276f62e80564d3cf5a92b9e6c04638364b5e39c