Set up packages for publishing to NPM
Summary:
The Hermes parser compiled to WASM is ready to be published as a `hermes-parser` NPM package. We are also working on a `hermes-eslint` package which we will also want to publish to NPM once an initial version is complete.

First of all let's restructure to have clear, separate packages. Under `hermes-parser/js` we now have two separate directories for the `hermes-parser` and `hermes-eslint` packages. Each of these package directories contains its own `package.json` which has been filled out, a `README.md` (will be created in the future for `hermes-eslint`), a `LICENSE` file containing a copy of the LICENSE at the root of the Hermes repo, all source code under the `src` directory, and all tests under the `__tests__` directory.

Both the `hermes-parser` and `hermes-eslint` packages are still built at the same time via `yarn build` in the `hermes-parser/js` directory. Builds are now created in each package directory within the more standard `dist` directory. The most notable change here is that we now run all source code (both generated and otherwise) through Babel to lower out all ES6 and experimental features. The lowered code after being run through Babel is what's written to the `dist` directory.

The `package.json` for each package specifies that the built `dist` files should be included in the final package instead of the `src` directory, and that the main for each package is `dist/index.js`. This means that running `yarn pack` in each package directory will bundle up all files in the `dist` directory, as well as the `package.json`, `README.md`, and `LICENSE` file which are all included by default. The result of `yarn pack` is ready to publish to NPM.

Reviewed By: avp

Differential Revision: D25830138

fbshipit-source-id: bb54ea7261dd223e36616c8f6756fa2689fa56d4