Remove CompleteMarkState from Hades dependencies
Summary:
CompleteMarkState accidentally got into the dependency list for Hades
via the `AlignedHeapSegment.h` header. Now that there's a specific header
for GenGC: `GenGCHeapSegment.h`, move the CompleteMarkState dependency
to there instead.
This means fewer files need to be compiled for Hades that are unused, and we
can remove an unused function on Hades: markSymbol.

Reviewed By: neildhar

Differential Revision: D25854223

fbshipit-source-id: 7ec62a77064375e49bba488ca3424ac68dfb9214