Change VM cells to have public constructors
Summary:
The GC needs to be a friend for `GC::makeA` to be able to construct
objects with a private constructor. This creates a lot of friend noise
in cells.
It's better to simply require the constructor to be public. The chance of
it being misused to construct JS cells on the malloc heap is minimal, since
they have all sorts of heap-specific stuff like write barriers on GCPointer members.

Reviewed By: tmikov

Differential Revision: D25507150

fbshipit-source-id: 2dbd0e2b90e3a72fe4d55163387f3299b05e9c92