new strategy for TDZ
Summary:
Previously we implemented TDZ by creating a shadow variable tracking the
validity of the "main" variable. The motivation was that tracking the
value of the shadow variable is easier and more optimizable.
Unfortunately that strategy doesn't work so well with dynamic runtime
variable resolution - we can't get information about the shadow variable
at runtime, and whether there even is a shadow variable at all.

So, we are switching to a new strategy, where TDZ is encoded by using
the "empty" HermesValue tag (which is also used to encode "holes" in
arrays). We emit the new instruction `ThrowIfEmpty` before using and TDZ
value.

Unlike `ThrowIfUndefined`, `ThrowIfEmpty` returns a result, enabling us
to use type inference to eliminate unnecessary TDZ checks (to be added
in a subsequent diff).

A couple of optimization-related tests are temporarily disabled.

Reviewed By: avp

Differential Revision: D25750411

fbshipit-source-id: ab4fb686ec1255489084e77ac48852aacedba20a