new bytecode: ThrowIfEmpty, LoadConstEmpty
Summary:
Add two new instructions to the interpreter. They are used in subsequent
diffs for a new approach to TDZ:

`ThrowIfEmpty r1, r2` - throw ReferenceError if r2 is the "empty"
HermesValue, otherwise move it to r1.

`LoadConstEmpty r1` - load the "empty" HermesValue into r1.

Note: the initial plan was for `ThrowIfEmpty` to have a single argument,
and to emit an extra `Mov` if the result register is different from the
input register, with the assumption that it would be a rare occurrence.
Unfortunately it turned out that the register allocator chooses a
different register too often. We will look into that later, but for now
it is a two operand instruction.

Reviewed By: neildhar

Differential Revision: D25750412

fbshipit-source-id: 720db7b3b947b89df5f2c70950b73398e9788399