Mark Interpreter::interpretFunction as noinline
Summary:
It should be illegal to inline `interpretFunction`, because it stores
label addresses in the `opcodeDispatch` local static variable.

This is because in the case where calls to `interpretFunction` are
inlined, we end up with multiple copies of the code for the function,
and, effectively, multiple copies of each label. However, we still
only have a single copy of the `opcodeDispatch` static member (since
it is static). Attempting to execute an inlined copy of the function
will access this shared copy, which contains pointers to the next
instruction to jump to, resulting in erratic control flow that jumps
back to the outlined copy of the function.

Due to a bug in handling this case LLVM, this function may
occasionally be inlined anyway.

To avoid this, explicitly mark the function as noinline until the
underlying bug is resolved.

Reviewed By: tmikov

Differential Revision: D25881099

fbshipit-source-id: b3ad8dddd18e51b714f4ccb88ed40012f8744d31