Require aligned sizes at GC::makeA
Summary:
In all of Hermes, we require every JS heap allocation to begin on an 8-byte
aligned address. We call the 8 byte alignment requirement the "heap alignment".

There was a bug in GenGC where `GenGC::growTo` was called with a
value that was not heap aligned, such as 19 bytes, and it grew by exactly
19 bytes.
Later, in `AlignedHeapSegment::alloc`, it was rounded up to a heap aligned size,
such as 24 bytes. Unfortunately the previous call to `growTo` didn't guarantee this
would succeed due to the alignment difference, so it would fail to allocate where we
asserted that it would be impossible.

To fix this specific issue and prevent any more similar issues from occurring across any
GC, add a requirement to any call to `GC::makeA` (the only allocation entrypoint) has to
use a heap aligned size. This will make all of the computation done on sizes in GenGC
work with only heap aligned sizes.
Usage of "external memory" must be careful to not produce a size that is not heap aligned.
Luckily in all of our GCs external memory is only ever compared to the amount needed for
a new allocation, so no issue comes up.

There is an alternative: have `makeA` call `heapAlignSize` on its value. This adds some
computation to the alloc path that didn't exist before, but guarantees we don't miss any
call sites in production. I think the precondition and assert are good enough to catch all
cases in Hermes (along with a thorough eye-based audit).

Reviewed By: neildhar

Differential Revision: D25964346

fbshipit-source-id: d4e31849fe4da8a64f7a72c59c7b7f3401c38897