Pull in fork of eslint-scope for ScopeManager
Summary:
The last component needed for a custom ESLint backend is a ScopeManager. Like `babel-eslint` and `typescript-eslint` we can base this off the existing [eslint-scope](https://github.com/eslint/eslint-scope) library, which implements the correct API needed for ESLint's scope manager.

Unlike `babel-eslint` we cannot just build on top of the `eslint-scope` package if we want to correctly maintain both value and type imports for Flow, as `eslint-scope` has no way of distinguishing between type and value imports. Instead we can fork `eslint-scope` (similar to `typescript-eslint` which is also based off a fork of `eslint-scope`), then augument with Flow specific features, as `eslint-scope` uses the permissive BSD-2 license.

This diff pulls in the latest version of the `eslint-scope` source code into a new `eslint-scope` directory within the `hermes-eslint` package. We make sure to properly preserve attribution by keeping the license headers on each file (as they differ from file to file), keeping a copy of the license in the `ESLINT_SCOPE_LICENSE` file, and including a README for the `eslint-scope` directory mentioning that it is forked with a link to the original project and a note about the license.

The only modifications made to the original `eslint-scope` source code is adding the `format` directive and formatting with prettier, as well as removing a handful of unnecessary vim setting and `eslint-enable/disable` comments.

The scope manager is not yet integrated with the rest of the `hermes-eslint` package. Future diffs will start to modify this library to manage the scope for Flow types.

Reviewed By: gkz

Differential Revision: D25964899

fbshipit-source-id: e3085712c31c89812a292bcda246ea7b4b74eff7