Pass allocations through GCBase
Summary:
Move the allocation tracking logic into GC code as well. This allows
the GC to make allocations without needing to call back to runtime
code, which avoids a circular dependency. More generally, in tests,
this allows us to write APIs that may potentially make allocations,
even when a `DummyRuntime` is used.

For example, for HV32, HermesValue functions will need the ability to
allocate, and it is easier to pass in a `GC *` than a `Runtime *`
since that preserves the ability to write tests with a `DummyRuntime`.

Reviewed By: dulinriley

Differential Revision: D25949779

fbshipit-source-id: b1d7a348693331fc6e741cc654c307fd013ba2df