Implement Rudimentary Async/Await
Summary:
This diff implements rudimentary async/await by lowering in the IRGen
async function (declarations) into an inner generator wrapped by a call
to the JS builtin function spawnAsync, with the await expr rewritten
as yield of the inner generator.

Support for async function expression, async arrow, and AsyncFunction
builtin object will be implemented in upper diffs.

N.B. non-simple parameters need to be took care so we are not emitting
duplicated code.

N.B. potential non-existence of ES6 Promise and ES6 Generator need to
be took care. In this diffs, we demonstrate:
- how the compiler error out when generator is disabled in APITest.
- how the VM would ill-behave when promise is not provided.
  - we will strength the VM check via bytcode header in upper diffs.

Reviewed By: tmikov, avp

Differential Revision: D24702359

fbshipit-source-id: e26fa42f8833a2d320fa023faa14484e30939608