Add option to list opcodes
Summary:
Add option to hbcdump's objdump format to also output an opcode list in the form of a disassembled synthetic function that has all opcodes and zeroed operand bytes. The synthetic function is distinguished by a special ID (0xffffffff).

This can be used by other tools to learn the basics about instructions that may not yet be present in the training bytecode bundle.

Reviewed By: avp

Differential Revision: D26056124

fbshipit-source-id: cd1b7526f223b9ee6c9a0f2d3557ea83cd15a46c