Support Flow enums in scope manager
Summary:
Let's add support for Flow enums to the new scope manager. We can create a new `EnumDefinition` class for enum name definitions, which can be referenced as both types and values. Then we simply need to add an enum definition to the current scope when we encounter an `EnumDeclaration` in the referencer.

Note that enum bodies may contain identifier nodes, but these are just member names and enum bodies cannot contain identifier references. This means we should not visit the enum's body in the referencer to avoid mistakenly counting enum member identifiers as references.

Reviewed By: gkz

Differential Revision: D26055202

fbshipit-source-id: d258ef36e34c434325cb25abb6bebf600ef80bd9