Support type parameters in scope manager
Summary:
Let's add support for Flow's type parameters to the new scope manager.

Type parameters are defined using the new `TypeParameterDefinition`, every time we encounter a `TypeParameter` node in the referencer. A complicating factor is that the ScopeManager API expects there to be an `Identifier` node that can be referenced for each definition, however `TypeParameter` nodes do not contain an `Identifier` node and instead have the name as a string under the `name` property. Since the ScopeManager API expects there to be an `Identifier` we must create a fake `Identifier` node with the correct location and name, which can be used if callers need an `Identifier` for `TypeParameter` nodes. Thankfully this doesn't negatively affect the rest of the ScopeManager, as the `Identifier` node isn't needed for anything other than its name.

This diff adds support for type parameters in all nodes that can contain type parameters (except for `declare` nodes which have not yet been implemented in the scope manager). For `TypeAlias`, `OpaqueType`, `InterfaceDeclaration`, and `FunctionTypeAnnotation` nodes we must also add a new `TypeScope` that contains the type parameters. Functions and classes will put the type parameters in the existing function or class scope, so we do not need to add a new `TypeScope` in these cases. We only add a type scope if there are any type parameters on a given node.

Let's also introduce a `ScopeType` object to provide a central definition for all scope types instead of using string literals at all use sites.

Reviewed By: gkz

Differential Revision: D26092800

fbshipit-source-id: 2b6fcf598dacb7f00c7f10fe3924440eb677d61d