Don't mix translated and untranslated mappings in the same input file
Summary:
When a CommonJS input file has a source map with unmapped regions, we currently emit mappings from the bytecode to the input file for those regions - just like we do for the entire file when there is *no* input source map available. This can result in a source map that mixes references to true source files and to intermediate compilation artifacts.

Here we refine this logic to instead emit unmapped regions (bytecode locations with no represented source locations), but only in the case that the input file *has* a source map in the first place. In that case the input file is by definition a compilation artifact so shouldn't be listed in the final source map.

Reviewed By: avp

Differential Revision: D26146134

fbshipit-source-id: 59afb88861e2531f387faf43e22b6b242fb9f857