Update scope manager to support type annotations in patterns
Summary:
Let's add support for visiting Flow type annotations within patterns to the scope manager. Patterns are currently visited by the `PatternVisitor`, which walks the pattern and calls a callback on each Identifier it finds. The pattern visitor also accumulates nodes that the referencer needs to visit, and optionally visits them depending on how the `visitPattern` helper function is called in the referencer.

We must update the pattern visitor to also add all Flow type annotation nodes to the collection of nodes to visit afterwards in the referencer. Let's also rename this collection of nodes from `rightHandNodes` to `extraNodesToVisit` for clarity, now that this also includes type annotations. We also can then rename the `processRightHandNodes` option passed around the referencer to the more accurate `visitAllNodes`.

Reviewed By: gkz

Differential Revision: D26135405

fbshipit-source-id: 9d0bee0cceb99486345fad2fd0de0566b374e2a2