Scope manager supports DeclareModule node and its exports
Summary:
Now that all other declare nodes have been implemented in the scope manager, let's implement the `DeclareModule` node as well. The `DeclareModule` node introduces its own `DeclareModuleScope` as names defined within the declare module body should not be in scope outside the declare module statement. Note that the scope manager will also place a `BlockScope` around the declare module body as well, but we still need the `DeclareModuleScope` wrapping that to prevent vars from escaping.

We make sure to not visit the "id" property of the declare module node, as the module name is not a definition that can be referenced (or a reference itself).

Now that we have support for the `DeclareModule` node, let's also add tests for all `declare export` cases.

Reviewed By: gkz

Differential Revision: D26139125

fbshipit-source-id: e5fcf05b086af46a9c5273cfe133c0c0d2b019f6