Consolidate Hades phase transitions in incrementalCollect
Summary:
This diff moves concurrentPhase_ updates into the incrementalCollect
function. It does the following:
1. Remove the WeakMapScan phase.
2. Move the advancement to the sweep phase, and waking up the
background thread, into incrementalCollect.

This makes the CompleteMarking phase a single monolithic phase
that covers the entirety of the STW pause, so we can add more things
to it without worrying about adding a new phase.

Reviewed By: dulinriley

Differential Revision: D26139403

fbshipit-source-id: 36b7aeb683c1d508fe573005eaf1e1f34a973cba