Disable FuncSigOpts for generators and async functions.
Summary:
FuncSigOpts was optimizing all analyzable calls to generators and
async functions and replacing all arguments with `undefined`.
It does this because it doesn't see the use of parameters in the
inner function when analyzing the callsites for the outer function,
and thinks they are all unused.

Disable FuncSigOpts for generators and async functions because
they should be considered to be users of all their parameters.

Reviewed By: tmikov

Differential Revision: D26243257

fbshipit-source-id: c2e04ba80c1748163d2db48e1fa655d3496a5244