Class property names are not treated as references by scope manager
Summary: The scope manager currently has a bug where it mistakenly treats class property names as references. This is simply because the `key` child of a `ClassProperty` node is naively treated as a reference when it is an identifier. Instead, similar to object properties we should only be visiting the `key` child if the property is computed (as the key is then an expression), otherwise the key will be an identifier or string literal that should not be treated as a reference.

Reviewed By: gkz

Differential Revision: D26255545

fbshipit-source-id: e8ae8db506f7bdab22bcc3d7d66d30f957116545