Fix mocking of HermesInternal.getInstrumentedStats
Summary:
The code to mock the output of HermesInternal.getInstrumentedStats in a synth trace had a bug where it used the decimal representation of numerical indices of predefined string keys. This kind-of-sort-of worked, until the numbering of those predefined strings changed. Fix it by using the actual string content instead.

This shouldn't break existing traces more than they already were.

Reviewed By: dulinriley

Differential Revision: D26287464

fbshipit-source-id: fe51973600bc8a23430daf30f391af4ce76c4319