Function expression type parameters and return type visited before function name in scope manager
Summary:
The order that the scope manager currently visits names in functions (and function expressions) is not quite correct. First of all type parameters should be in a scope that contains the function expression name scope, meaning the function expression name will shadow a type parameter with the same name in the body of the function expression. However the function expression name is not in scope for the function's return type (though the function's type parameters are in scope).

This means that we must add a type scope encompassing the function expression name and function scope if type parameters exist. We can then define the type parameters here and then visit the return type before creating the inner function scopes. This is the same logic and traversal order used by `babel-eslint`.

Reviewed By: gkz

Differential Revision: D26280330

fbshipit-source-id: 8fb4a3a28ebb4f5dae259c12b779def094b4a9bf