Add workspaces and prepare for publishing to NPM
Summary:
An initial version of the `hermes-eslint` package is complete - let's get it ready for publishing to NPM.

Now that we have multiple packages that depend on each other we should be using workspaces so that the dependency is resolved locally if the versions match. We also now need to actually specify versions so that `hermes-eslint` can depend on the latest version of `hermes-parser`. We can set this to `0.4.0` as this is the next version that will be released, which will be published once this diff is out.

Now that we are using workspaces we no longer need to `yarn install` for each package - a `yarn install` in the root workspace will install all packages in the workspace.

Let's also add a basic README mentioning how to use `hermes-eslint` and what options are allowed.

Reviewed By: pieterv

Differential Revision: D26291161

fbshipit-source-id: aae927c0716cfd3cfbea4f924ddf9bfc5ce17d0d