Fix lexing for fractional numbers with leading 0.
Summary:
Non-integer numbers with leading zeroes should be switched to decimal
if they contain `8` or `9` in the integer part.
If they don't contain either of those, they should error because legacy
octal literals are supposed to be integers.

For example, in sloppy mode, both `08.11` and `08.9` are valid,
but `07.11` is not valid.

Any non-integer number with a leading zero should error in strict mode.

Factor out the `8` and `9` detection code and make it resilient when
`real` is true.

Reviewed By: tmikov

Differential Revision: D25625741

fbshipit-source-id: 4addfa78dfae0cacd3e953a738bfa03829fbb827