Support this type annotations on functions in hermes-parser and hermes-eslint
Summary:
Now that Hermes can parse Flow's new this type annotations the WASM parser and eslint utilities should be updated to work with this type annotations.

This type annotations on function declarations and function expressions are added as the first parameter in the AST. When then WASM parser is in ESTree mode this is okay, as we control the ESTree AST format for Flow nodes. However when the WASM parser is instead producing a Babel AST we must strip out these this type annotations from the params list, as Babel does not recognize this AST structure.

The `hermes-eslint` package uses the ESTree AST so it will receive this type annotations. However the scope manager should not treat this type annotations on functions as actual parameter definitions that can be referenced. We can simply prevent definitions from being created for the first function parameter if it is named "this".

Reviewed By: gkz

Differential Revision: D26358212

fbshipit-source-id: 1e8c44bf85db72c4621713a267e68eeee6f8d92c