optionally record a trace of the last 8 instructions
Summary:
The "crash trace" records:
- Segment ID, source URL and source hash of the runtime module of the
latest bytecode function
- File-relative offsets and the opcodes of the last eight instructions.

To compile in the new functionality, there is a new build mode
controlled by the preprocessor define `HERMESVM_CRASH_TRACE`. It must be
set to `1` to enable it (because we check the value in C++ code).

In CMake there is an option with the same name, in Buck the mode is
called `hermes.crash_trace`.

Once compiled in, at runtime it is controlled by
`vm::experiments::CrashTrace = 1 << 13`. From the CLI:

    hermes -Xvm-experiment-flags=8192 file.hbc

When enabled, execution time increased by about 45%.

Reviewed By: neildhar

Differential Revision: D26323871

fbshipit-source-id: c203046eefd05f9643d3e6d69a8fb9198d511e66